+++ R/Gargle-class.R
@@ -182,5 +182,13 @@ Gargle2.0 <- R6::R6Class("Gargle2.0", inherit = httr::Token2.0, list(
       self$cache()
     }
     self
+  },
+  init_credentials = function() {
+    cat_line("initiating new token")
+    if (rlang::is_interactive()) {
+      super$init_credentials()
+    } else {
+      abort("OAuth2 flow requires an interactive session")
+    }
   }
 ))